Replaced all long doubles with doubles. Now, I get full points. Funny how that works.

+++ Functions.cpp
@@ -9,121 +9,121 @@ funcPtr Functions::derivatives[funcTypeLen];
 std::map<std::string, FuncType> Functions::funcNames;
 
 // one parameter functions
-long double Functions::getCos(Variable *a, Variable *b) {
+double Functions::getCos(Variable *a, Variable *b) {
   return cos(a -> getComputedValue());
 }
 
-long double Functions::getSin(Variable *a, Variable *b) {
+double Functions::getSin(Variable *a, Variable *b) {
   return sin(a -> getComputedValue());
 }
 
-long double Functions::getTan(Variable *a, Variable *b) {
+double Functions::getTan(Variable *a, Variable *b) {
   return tan(a -> getComputedValue());
 }
 
-long double Functions::getACos(Variable *a, Variable *b) {
+double Functions::getACos(Variable *a, Variable *b) {
   return acos(a -> getComputedValue());
 }
 
-long double Functions::getASin(Variable *a, Variable *b) {
+double Functions::getASin(Variable *a, Variable *b) {
   return asin(a -> getComputedValue());
 }
 
-long double Functions::getATan(Variable *a, Variable *b) {
+double Functions::getATan(Variable *a, Variable *b) {
   return atan(a -> getComputedValue());
 }
 
-long double Functions::getExp(Variable *a, Variable *b) {
+double Functions::getExp(Variable *a, Variable *b) {
   return exp(a -> getComputedValue());
 }
 
-long double Functions::getLog(Variable *a, Variable *b) {
+double Functions::getLog(Variable *a, Variable *b) {
   return log(a -> getComputedValue());
 }
 
-long double Functions::getSqrt(Variable *a, Variable *b) {
+double Functions::getSqrt(Variable *a, Variable *b) {
   return sqrt(a -> getComputedValue());
 }
 
 // one parameter derivative functions
-long double Functions::deriveCos(Variable *a, Variable *b) {
+double Functions::deriveCos(Variable *a, Variable *b) {
   return -getSin(a) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveSin(Variable *a, Variable *b) {
+double Functions::deriveSin(Variable *a, Variable *b) {
   return getCos(a) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveTan(Variable *a, Variable *b) {
-  return 1.0L / pow(getCos(a), 2) * (a -> getDerivativeValue());
+double Functions::deriveTan(Variable *a, Variable *b) {
+  return 1.0 / pow(getCos(a), 2) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveACos(Variable *a, Variable *b) {
-  return (-1.0L / sqrt(1.0L - pow(a -> getComputedValue(), 2))) * (a -> getDerivativeValue());
+double Functions::deriveACos(Variable *a, Variable *b) {
+  return (-1.0 / sqrt(1.0 - pow(a -> getComputedValue(), 2))) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveASin(Variable *a, Variable *b) {
-  return (1.0L / sqrt(1.0L - pow(a -> getComputedValue(), 2))) * (a -> getDerivativeValue());
+double Functions::deriveASin(Variable *a, Variable *b) {
+  return (1.0 / sqrt(1.0 - pow(a -> getComputedValue(), 2))) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveATan(Variable *a, Variable *b) {
-  return (1.0L / (1.0L + pow(a -> getComputedValue(), 2))) * (a -> getDerivativeValue());
+double Functions::deriveATan(Variable *a, Variable *b) {
+  return (1.0 / (1.0 + pow(a -> getComputedValue(), 2))) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveExp(Variable *a, Variable *b) {
+double Functions::deriveExp(Variable *a, Variable *b) {
   return getExp(a) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveLog(Variable *a, Variable *b) {
-  return (1.0L / (a -> getComputedValue())) * (a -> getDerivativeValue());
+double Functions::deriveLog(Variable *a, Variable *b) {
+  return (1.0 / (a -> getComputedValue())) * (a -> getDerivativeValue());
 }
 
-long double Functions::deriveSqrt(Variable *a, Variable *b) {
-  return (1.0L / (2.0L * getSqrt(a))) * (a -> getDerivativeValue());
+double Functions::deriveSqrt(Variable *a, Variable *b) {
+  return (1.0 / (2.0 * getSqrt(a))) * (a -> getDerivativeValue());
 }
 
 // two parameter functions
-long double Functions::getMult(Variable *a, Variable *b) {
+double Functions::getMult(Variable *a, Variable *b) {
   return (a -> getComputedValue()) * (b -> getComputedValue());
 }
 
-long double Functions::getAdd(Variable *a, Variable *b) {
+double Functions::getAdd(Variable *a, Variable *b) {
   return (a -> getComputedValue()) + (b -> getComputedValue());
 }
 
-long double Functions::getSubs(Variable *a, Variable *b) {
+double Functions::getSubs(Variable *a, Variable *b) {
   return (a -> getComputedValue()) - (b -> getComputedValue());
 }
 
-long double Functions::getDivide(Variable *a, Variable *b) {
+double Functions::getDivide(Variable *a, Variable *b) {
   return (a -> getComputedValue()) / (b -> getComputedValue());
 }
 
-long double Functions::getPow(Variable *a, Variable *b) {
+double Functions::getPow(Variable *a, Variable *b) {
   return pow(a -> getComputedValue(), (b -> getComputedValue()));
 }
 
 // two parameter derivates
-long double Functions::deriveMult(Variable *a, Variable *b) {
+double Functions::deriveMult(Variable *a, Variable *b) {
   return (a -> getDerivativeValue()) * (b -> getComputedValue()) +
          (a -> getComputedValue()) * (b -> getDerivativeValue());
 }
 
-long double Functions::deriveAdd(Variable *a, Variable *b) {
+double Functions::deriveAdd(Variable *a, Variable *b) {
   return (a -> getDerivativeValue()) + (b -> getDerivativeValue());
 }
 
-long double Functions::deriveSubs(Variable *a, Variable *b) {
+double Functions::deriveSubs(Variable *a, Variable *b) {
   return (a -> getDerivativeValue()) - (b -> getDerivativeValue());
 }
 
-long double Functions::deriveDivide(Variable *a, Variable *b) {
+double Functions::deriveDivide(Variable *a, Variable *b) {
   return ((a -> getDerivativeValue()) * (b -> getComputedValue()) -
           (a -> getComputedValue()) * (b -> getDerivativeValue())) /
          pow(b -> getComputedValue(), 2);
 }
 
-long double Functions::derivePow(Variable *a, Variable *b) {
+double Functions::derivePow(Variable *a, Variable *b) {
   // includes cases such as x^x => (a^b)' = e^bloga * (b'loga + ba'/a)
   return exp(b -> getComputedValue() * getLog(a)) * ((b -> getDerivativeValue()) * getLog(a) +
                                                      (b -> getComputedValue()) * (a -> getDerivativeValue()) /
